requirements.txt
# Core
python-telegram-bot==20.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2

# Text embeddings + clustering
bertopic==0.16.0
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
torchvision==0.17.1+cpu
--find-links https://download.pytorch.org/whl/torch_stable.html

# Utils
umap-learn==0.5.5
hdbscan==0.8.33
tqdm==4.66.5
rich==13.7.1

# Для обработки русского языка
pymorphy2
pymorphy2-dicts-ru


